Move ServiceMethod to psibase library
